requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
blinker==1.9.0
bootstrap4==0.1.0
boto3==1.41.5
botocore==1.41.5
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dash==3.2.0
dash-bootstrap-components==2.0.4
dash_mantine_components==2.3.0
Django==5.2.6
django-bootstrap5==25.2
django-pgviews==0.6.0
django-plotly-dash==2.5.0
dpd-static-support==0.0.5
dpd_components==0.2.0
Flask==3.1.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
plotly==6.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
retrying==1.4.2
s3transfer==0.15.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.5
urllib3==2.6.3
whitenoise==6.11.0
zipp==3.23.0
zope.interface==8.0.1